Revert "refactor(tools): score based diff merges (#1494)"
This reverts commit c23eea6121aa32050e2557d51bd433b340f68308.
