--- tests/fixtures/files-diff-1.3.patch
@@ -1,12 +1,0 @@
-*** Begin Patch
-@@ two
-@@ three
-four
--five
-+5
-six
-
-@@ seven
-eight
--nine
-*** End Patch
