--- tests/fixtures/files-diff-1.4.patch
@@ -1,15 +1,0 @@
-*** Begin Patch
-@@ two
-three
-four
--five
-+5
-six
-seven
-*** End Patch
-
-*** Begin Patch
-seven
-eight
--nine
-*** End Patch
