--- tests/fixtures/files-diff-1.5.patch
@@ -1,10 +1,0 @@
-*** Begin Patch
-@@ two
-three
-four
-five
--five
-+5
-six
-seven
-*** End Patch
