+++ tests/fixtures/files-diff-1.patch
