--- tests/fixtures/files-diff-2.1.patch
@@ -1,12 +1,0 @@
-*** Begin Patch
-@@ 2nd:
-one
-two
--three
-+3
-four
-five
--six
-+6
-seven
-*** End Patch
