+++ tests/fixtures/files-diff-2.patch
