+++ tests/fixtures/files-diff-4.2.patch
