+++ tests/fixtures/files-diff-4.patch
@@ -1,19 +1,12 @@
 *** Begin Patch
-@@    <article>
-
-      <header>
-        <h1>Deeply Nested Structure</h1>
--        <br>
-      </header>
-
-      <main>
-
-@@    <article>
-      <footer>
-        <p>End of Article</p>
--        <br>
-+        <p>Good work</p>
-      </footer>
-
-    </article>
+@@ 2nd:
+one
+two
+-three
++3
+four
+five
+-six
++6
+seven
 *** End Patch
