+++ tests/fixtures/files-diff-5.patch
@@ -1,18 +1,7 @@
 *** Begin Patch
-@@
--local function greet(name)
-+---@param name string|nil
-+---@return string
-+local function greet(name)
-  name = name or "world"
-  local message = "Hello, " .. name .. "!"
-  print(message)
-  return message
-@@end
-
-+---@param tbl table
- ---@return number
- local function count(tbl)
-   if type(tbl) ~= "table" then
-     error("count: expected table, got " .. type(tbl))
+@@ <div id="l1">
+<li>
+-  <span><em>Bee</em></span>
++  <span><em>B</em></span>
+</li>
 *** End Patch
