+++ tests/fixtures/files-diff-6.patch
@@ -1,0 +1,19 @@
+*** Begin Patch
+@@    <article>
+
+      <header>
+        <h1>Deeply Nested Structure</h1>
+-        <br>
+      </header>
+
+      <main>
+
+@@    <article>
+      <footer>
+        <p>End of Article</p>
+-        <br>
++        <p>Good work</p>
+      </footer>
+
+    </article>
+*** End Patch
