--- tests/fixtures/files-input-2.html
@@ -1,35 +1,0 @@
-1st:
-one
-two
-three
-four
-five
-six
-seven
-eight
-nine
-ten
-
-2nd:
-one
-two
-three
-four
-five
-six
-seven
-eight
-nine
-ten
-
-3nd:
-one
-two
-three
-four
-five
-six
-seven
-eight
-nine
-ten
