+++ tests/fixtures/files-input-4.html
@@ -1,66 +1,35 @@
-<!DOCTYPE html>
-<html>
+1st:
+one
+two
+three
+four
+five
+six
+seven
+eight
+nine
+ten
 
-<head>
-  <title>Nested HTML with Line Breaks</title>
-</head>
+2nd:
+one
+two
+three
+four
+five
+six
+seven
+eight
+nine
+ten
 
-<body>
-
-<div id="root">
-
-  <section>
-
-    <article>
-
-      <header>
-        <h1>Deeply Nested Structure</h1>
-        <br>
-      </header>
-
-      <main>
-
-        <ul>
-
-          <li>
-            <span>
-              <em>Item 1</em>
-              <br><br>
-            </span>
-          </li>
-
-          <li>
-            <span>
-              <em>Item 2</em>
-              <br>
-              <br>
-              <br>
-            </span>
-          </li>
-
-          <li>
-            <span>
-              <em>Item 3</em>
-              <br>
-            </span>
-          </li>
-
-        </ul>
-
-      </main>
-
-      <footer>
-        <p>End of Article</p>
-        <br>
-      </footer>
-
-    </article>
-
-  </section>
-
-</div>
-
-</body>
-
-</html>
-
+3nd:
+one
+two
+three
+four
+five
+six
+seven
+eight
+nine
+ten
