+++ tests/fixtures/files-input-6.html
@@ -1,0 +1,66 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <title>Nested HTML with Line Breaks</title>
+</head>
+
+<body>
+
+<div id="root">
+
+  <section>
+
+    <article>
+
+      <header>
+        <h1>Deeply Nested Structure</h1>
+        <br>
+      </header>
+
+      <main>
+
+        <ul>
+
+          <li>
+            <span>
+              <em>Item 1</em>
+              <br><br>
+            </span>
+          </li>
+
+          <li>
+            <span>
+              <em>Item 2</em>
+              <br>
+              <br>
+              <br>
+            </span>
+          </li>
+
+          <li>
+            <span>
+              <em>Item 3</em>
+              <br>
+            </span>
+          </li>
+
+        </ul>
+
+      </main>
+
+      <footer>
+        <p>End of Article</p>
+        <br>
+      </footer>
+
+    </article>
+
+  </section>
+
+</div>
+
+</body>
+
+</html>
+
