--- tests/fixtures/files-output-1.4.html
@@ -1,9 +1,0 @@
-one
-two
-three
-four
-5
-six
-seven
-eight
-ten
