--- tests/fixtures/files-output-1.5.html
@@ -1,10 +1,0 @@
-one
-two
-three
-four
-5
-six
-seven
-eight
-nine
-ten
