--- tests/fixtures/files-output-2.1.html
@@ -1,35 +1,0 @@
-1st:
-one
-two
-three
-four
-five
-six
-seven
-eight
-nine
-ten
-
-2nd:
-one
-two
-3
-four
-five
-6
-seven
-eight
-nine
-ten
-
-3nd:
-one
-two
-three
-four
-five
-six
-seven
-eight
-nine
-ten
