+++ tests/fixtures/files-output-3.html
@@ -1,30 +1,9 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Nested HTML Test</title>
-  </head>
-  <body>
-    <div id="l1">
-      <section>
-        <article>
-          <header>
-            <h1>H</h1>
-          </header>
-          <main>
-            <ul>
-              <li>
-                <span><em>A</em></span>
-              </li>
-              <li>
-                <span><em>B</em></span>
-              </li>
-              <li>
-                <span><em>C</em></span>
-              </li>
-            </ul>
-          </main>
-        </article>
-      </section>
-    </div>
-  </body>
-</html>
+one
+two
+three
+four
+5
+six
+seven
+eight
+ten
