+++ tests/fixtures/files-output-4.html
@@ -1,65 +1,35 @@
-<!DOCTYPE html>
-<html>
+1st:
+one
+two
+three
+four
+five
+six
+seven
+eight
+nine
+ten
 
-<head>
-  <title>Nested HTML with Line Breaks</title>
-</head>
+2nd:
+one
+two
+3
+four
+five
+6
+seven
+eight
+nine
+ten
 
-<body>
-
-<div id="root">
-
-  <section>
-
-    <article>
-
-      <header>
-        <h1>Deeply Nested Structure</h1>
-      </header>
-
-      <main>
-
-        <ul>
-
-          <li>
-            <span>
-              <em>Item 1</em>
-              <br><br>
-            </span>
-          </li>
-
-          <li>
-            <span>
-              <em>Item 2</em>
-              <br>
-              <br>
-              <br>
-            </span>
-          </li>
-
-          <li>
-            <span>
-              <em>Item 3</em>
-              <br>
-            </span>
-          </li>
-
-        </ul>
-
-      </main>
-
-      <footer>
-        <p>End of Article</p>
-        <p>Good work</p>
-      </footer>
-
-    </article>
-
-  </section>
-
-</div>
-
-</body>
-
-</html>
-
+3nd:
+one
+two
+three
+four
+five
+six
+seven
+eight
+nine
+ten
