+++ tests/fixtures/files-output-5.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Nested HTML Test</title>
+  </head>
+  <body>
+    <div id="l1">
+      <section>
+        <article>
+          <header>
+            <h1>H</h1>
+          </header>
+          <main>
+            <ul>
+              <li>
+                <span><em>A</em></span>
+              </li>
+              <li>
+                <span><em>B</em></span>
+              </li>
+              <li>
+                <span><em>C</em></span>
+              </li>
+            </ul>
+          </main>
+        </article>
+      </section>
+    </div>
+  </body>
+</html>
