+++ tests/fixtures/files-output-6.html
@@ -1,0 +1,65 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <title>Nested HTML with Line Breaks</title>
+</head>
+
+<body>
+
+<div id="root">
+
+  <section>
+
+    <article>
+
+      <header>
+        <h1>Deeply Nested Structure</h1>
+      </header>
+
+      <main>
+
+        <ul>
+
+          <li>
+            <span>
+              <em>Item 1</em>
+              <br><br>
+            </span>
+          </li>
+
+          <li>
+            <span>
+              <em>Item 2</em>
+              <br>
+              <br>
+              <br>
+            </span>
+          </li>
+
+          <li>
+            <span>
+              <em>Item 3</em>
+              <br>
+            </span>
+          </li>
+
+        </ul>
+
+      </main>
+
+      <footer>
+        <p>End of Article</p>
+        <p>Good work</p>
+      </footer>
+
+    </article>
+
+  </section>
+
+</div>
+
+</body>
+
+</html>
+
