+++ tests/fixtures/files-output-7.lua
